+++ source/commands/inventory.js
@@ -1,6 +1,5 @@
 const { PermissionFlagsBits, AttachmentBuilder, InteractionContextType, MessageFlags } = require('discord.js');
 const { CommandWrapper } = require('../classes/index.js');
-const { Op } = require('sequelize');
 const { commandMention } = require('../util/textUtil.js');
 const { MAX_MESSAGE_CONTENT_LENGTH } = require('../constants.js');
 
@@ -10,7 +9,7 @@ let logicLayer;
 const mainId = "inventory";
 module.exports = new CommandWrapper(mainId, "Show your inventory of usable items", PermissionFlagsBits.ViewChannel, false, [InteractionContextType.BotDM, InteractionContextType.Guild, InteractionContextType.PrivateChannel], 3000,
 	(interaction, database, runMode) => {
-		database.models.Item.findAll({ where: { userId: interaction.user.id, count: { [Op.gt]: 0 } } }).then(itemRows => {
+		logicLayer.items.getInventory(interaction.user.id).then(itemRows => {
 			let content = `Here are the items in your inventory (use them with ${commandMention("item")}):\n- `;
 			if (itemRows.length > 0) {
 				content += `${itemRows.map(row => `${row.itemName} x ${row.count}`).join("\n- ")}`;