+++ source/context_menus/Raise_a_Toast.js
@@ -93,7 +93,7 @@ module.exports = new UserContextMenuWrapper(mainId, PermissionFlagsBits.SendMess
 							thread.send({ content, flags: MessageFlags.SuppressNotifications });
 						})
 					}
-					updateScoreboard(company, modalSubmission.guild, database, logicLayer);
+					updateScoreboard(modalSubmission.guild, database, logicLayer);
 				}
 			});
 		})