+++ source/items/profile-colorizer-darkpurple.js
@@ -4,15 +4,12 @@ const { Item } = require("../classes");
 /** @type {typeof import("../logic")} */
 let logicLayer;
 
-const color = "Blue";
+const color = "Dark Purple";
 const itemName = `${color} Profile Colorizer`;
 module.exports = new Item(itemName, `Changes the color of your stats profile embed to ${color.toLowerCase()}`, 3000,
 	/** Sets the user's Hunter profile to Colors.Blue in the used guild */
 	async (interaction, database) => {
-		logicLayer.hunters.findOneHunter(interaction.user.id, interaction.guild.id).then(hunter => {
-			hunter.profileColor = color;
-			hunter.save();
-		})
+		logicLayer.hunters.setHunterProfileColor(interaction.user.id, interaction.guild.id, color.replace(/ /g, ""));
 		interaction.reply({ content: `Your profile color has been set to ${color} in this server.`, flags: [MessageFlags.Ephemeral] });
 	}
 ).setLogicLinker(logicBlob => {