+++ source/util/embedUtil.js
@@ -1,13 +1,8 @@
 const fs = require("fs");
 const { EmbedBuilder, Guild, Colors } = require("discord.js");
 const { Sequelize } = require("sequelize");
-const { Hunter } = require("../models/users/Hunter");
-const { Company } = require("../models/companies/Company");
-const { COMPANY_XP_COEFFICIENT, MAX_EMBED_DESCRIPTION_LENGTH } = require("../constants");
+const { MAX_EMBED_DESCRIPTION_LENGTH } = require("../constants");
 const { generateTextBar } = require("./textUtil");
-const { findLatestGoalProgress } = require("../logic/goals");
-const { findOrCreateCompany } = require("../logic/companies");
-const { findOrCreateCurrentSeason, findOneSeason } = require("../logic/seasons");
 
 const discordIconURL = "https://cdn.discordapp.com/attachments/618523876187570187/1110265047516721333/discord-mark-blue.png";
 const bountyBotIcon = "https://cdn.discordapp.com/attachments/618523876187570187/1138968614364528791/BountyBotIcon.jpg";
@@ -47,46 +42,13 @@ function randomFooterTip() {
 	return tipPool[Math.floor(Math.random() * tipPool.length)];
 }
 
-/**
- * @param {Guild} guild
- * @param {Sequelize} database
- */
-async function buildCompanyStatsEmbed(guild, database) {
-	const [company] = await findOrCreateCompany(guild.id);
-	const [currentSeason] = await findOrCreateCurrentSeason(guild.id);
-	const lastSeason = await findOneSeason(guild.id, "previous");
-	const participantCount = await database.models.Participation.count({ where: { seasonId: currentSeason.id } });
-	const companyXP = await company.xp;
-	const currentSeasonXP = await currentSeason.totalXP;
-	const lastSeasonXP = await lastSeason?.totalXP ?? 0;
-
-	const currentLevelThreshold = Hunter.xpThreshold(company.level, COMPANY_XP_COEFFICIENT);
-	const nextLevelThreshold = Hunter.xpThreshold(company.level + 1, COMPANY_XP_COEFFICIENT);
-	const particpantPercentage = participantCount / guild.memberCount * 100;
-	const seasonXPDifference = currentSeasonXP - lastSeasonXP;
-	const seasonBountyDifference = currentSeason.bountiesCompleted - (lastSeason?.bountiesCompleted ?? 0);
-	const seasonToastDifference = currentSeason.toastsRaised - (lastSeason?.toastsRaised ?? 0);
-	return new EmbedBuilder().setColor(Colors.Blurple)
-		.setAuthor(module.exports.ihpAuthorPayload)
-		.setTitle(`${guild.name} is __Level ${company.level}__`)
-		.setThumbnail(guild.iconURL())
-		.setDescription(`${generateTextBar(companyXP - currentLevelThreshold, nextLevelThreshold - currentLevelThreshold, 11)}*Next Level:* ${nextLevelThreshold - companyXP} Bounty Hunter Levels`)
-		.addFields(
-			{ name: "Total Bounty Hunter Level", value: `${companyXP} level${companyXP == 1 ? "" : "s"}`, inline: true },
-			{ name: "Participation", value: `${participantCount} server members have interacted with BountyBot this season (${particpantPercentage.toPrecision(3)}% of server members)` },
-			{ name: `${currentSeasonXP} XP Earned Total (${seasonXPDifference === 0 ? "same as last season" : `${seasonXPDifference > 0 ? `+${seasonXPDifference} more XP` : `${seasonXPDifference * -1} fewer XP`} than last season`})`, value: `${currentSeason.bountiesCompleted} bounties (${seasonBountyDifference === 0 ? "same as last season" : `${seasonBountyDifference > 0 ? `**+${seasonBountyDifference} more bounties**` : `**${seasonBountyDifference * -1} fewer bounties**`} than last season`})\n${currentSeason.toastsRaised} toasts (${seasonToastDifference === 0 ? "same as last season" : `${seasonToastDifference > 0 ? `**+${seasonToastDifference} more toasts**` : `**${seasonToastDifference * -1} fewer toasts**`} than last season`})` }
-		)
-		.setFooter(randomFooterTip())
-		.setTimestamp()
-}
-
 /** A seasonal scoreboard orders a company's hunters by their seasonal xp
  * @param {Guild} guild
  * @param {Sequelize} database
  */
 async function buildSeasonalScoreboardEmbed(guild, database, logicLayer) {
-	const [company] = await findOrCreateCompany(guild.id);
-	const [season] = await findOrCreateCurrentSeason(company.id);
+	const [company] = await logicLayer.companies.findOrCreateCompany(guild.id);
+	const [season] = await logicLayer.seasons.findOrCreateCurrentSeason(company.id);
 	const participations = await database.models.Participation.findAll({ where: { seasonId: season.id }, order: [["xp", "DESC"]] });
 
 	const hunterMembers = await guild.members.fetch({ user: participations.map(participation => participation.userId) });
@@ -124,7 +86,7 @@ async function buildSeasonalScoreboardEmbed(guild, database, logicLayer) {
 	}
 
 	const fields = [];
-	const { goalId, currentGP, requiredGP } = await findLatestGoalProgress(guild.id);
+	const { goalId, currentGP, requiredGP } = await logicLayer.goals.findLatestGoalProgress(guild.id);
 	if (goalId !== null) {
 		fields.push({ name: "Server Goal", value: `${generateTextBar(currentGP, requiredGP, 15)} ${currentGP}/${requiredGP} GP` });
 	}
@@ -147,7 +109,7 @@ async function buildSeasonalScoreboardEmbed(guild, database, logicLayer) {
  */
 async function buildOverallScoreboardEmbed(guild, database, logicLayer) {
 	const hunters = await database.models.Hunter.findAll({ where: { companyId: guild.id }, order: [["xp", "DESC"]] });
-	const [company] = await findOrCreateCompany(guild.id);
+	const [company] = await logicLayer.companies.findOrCreateCompany(guild.id);
 
 	const hunterMembers = await guild.members.fetch({ user: hunters.map(hunter => hunter.userId) });
 	const rankmojiArray = (await logicLayer.ranks.findAllRanks(guild.id, "descending")).map(rank => rank.rankmoji);
@@ -183,7 +145,7 @@ async function buildOverallScoreboardEmbed(guild, database, logicLayer) {
 	}
 
 	const fields = [];
-	const { goalId, currentGP, requiredGP } = await findLatestGoalProgress(guild.id);
+	const { goalId, currentGP, requiredGP } = await logicLayer.goals.findLatestGoalProgress(guild.id);
 	if (goalId !== null) {
 		fields.push({ name: "Server Goal", value: `${generateTextBar(currentGP, requiredGP, 15)} ${currentGP}/${requiredGP} GP` });
 	}
@@ -199,40 +161,6 @@ async function buildOverallScoreboardEmbed(guild, database, logicLayer) {
 	}
 
 	return embed;
-}
-
-/**
- * @param {Guild} guild
- * @param {GuildMember} member
- * @param {Hunter} hunter
- * @param {Sequelize} database
- */
-async function buildModStatsEmbed(guild, member, hunter, database) {
-	const embed = new EmbedBuilder().setColor(member.displayColor)
-		.setAuthor({ name: guild.name, iconURL: guild.iconURL() })
-		.setTitle(`Moderation Stats: ${member.user.tag}`)
-		.setThumbnail(member.user.avatarURL())
-		.setDescription(`Display Name: **${member.displayName}** (id: *${member.id}*)\nAccount created on: ${member.user.createdAt.toDateString()}\nJoined server on: ${member.joinedAt.toDateString()}`)
-		.addFields(
-			{ name: "Bans", value: `Currently Banned: ${hunter.isBanned ? "Yes" : "No"}\nHas Been Banned: ${hunter.hasBeenBanned ? "Yes" : "No"}`, inline: true },
-			{ name: "Disqualifications", value: `${await database.models.Participation.sum("dqCount", { where: { companyId: guild.id, userId: member.id } }) ?? 0} season DQs`, inline: true },
-			{ name: "Penalties", value: `${hunter.penaltyCount} penalties (${hunter.penaltyPointTotal} points total)`, inline: true }
-		)
-		.setFooter(randomFooterTip())
-		.setTimestamp();
-
-	let bountyHistory = "";
-	const lastFiveBounties = await database.models.Bounty.findAll({ where: { userId: member.id, companyId: guild.id, state: "completed" }, order: [["completedAt", "DESC"]], limit: 5 });
-	lastFiveBounties.forEach(async bounty => {
-		const completions = await database.models.Completion.findAll({ where: { bountyId: bounty.id } });
-		bountyHistory += `__${bounty.title}__${bounty.description !== null ? ` ${bounty.description}` : ""}\n${bounty.xpAwarded} XP per completer\nCompleters: <@${completions.map(completion => completion.userId).join('>, <@')
-			}>\n\n`;
-	})
-
-	if (bountyHistory === "") {
-		bountyHistory = "No recent bounties";
-	}
-	return embed.addFields({ name: "Last 5 Completed Bounties Created by this User", value: bountyHistory });
 }
 
 /** The version embed lists the following: changes in the most recent update, known issues in the most recent update, and links to support the project
@@ -260,11 +188,11 @@ async function buildVersionEmbed() {
 }
 
 /** If the guild has a scoreboard reference channel, update the embed in it
- * @param {Company} company
  * @param {Guild} guild
  * @param {Sequelize} database
  */
-function updateScoreboard(company, guild, database, logicLayer) {
+async function updateScoreboard(guild, database, logicLayer) {
+	const [company] = await logicLayer.companies.findOrCreateCompany(guild.id);
 	if (company.scoreboardChannelId && company.scoreboardMessageId) {
 		guild.channels.fetch(company.scoreboardChannelId).then(scoreboard => {
 			return scoreboard.messages.fetch(company.scoreboardMessageId);
@@ -277,10 +205,8 @@ function updateScoreboard(company, guild, database, logicLayer) {
 module.exports = {
 	ihpAuthorPayload: { name: "Click here to check out the Imaginary Horizons GitHub", iconURL: "https://images-ext-2.discordapp.net/external/8DllSg9z_nF3zpNliVC3_Q8nQNu9J6Gs0xDHP_YthRE/https/cdn.discordapp.com/icons/353575133157392385/c78041f52e8d6af98fb16b8eb55b849a.png", url: "https://github.com/Imaginary-Horizons-Productions" },
 	randomFooterTip,
-	buildCompanyStatsEmbed,
 	buildSeasonalScoreboardEmbed,
 	buildOverallScoreboardEmbed,
-	buildModStatsEmbed,
 	buildVersionEmbed,
 	updateScoreboard
 };