+++ ChangeLog.md
@@ -11,6 +11,8 @@ Server Goals are BountyBot objectives everyone on the server contributes to. A g
 - Added an optional text input for `/evergreen showcase` to allow moderators to add text to the showcase message
 - Fixed `/bounty edit` and `/evergreen edit` not saving modifications
 - Fixed crash when using `/create-default rank-roles` twice in a row
+- Fixed `/moderation user-report` listing XP awarded for recent bounties as "undefined"
+- Fixed Dark Purple Profile Colorizer not dropping
 ## BountyBot Version 2.8.0:
 ### Context Menu Options
 - Added the following BountyBot functionality to the Apps dropdown when right-clicking on a User or Message: